Fianlizando API de autores

+++ src/main/java/com/introducao/spring_jpa/service/AutorService.java
@@ -1,7 +1,11 @@
 package com.introducao.spring_jpa.service;
 
+import com.introducao.spring_jpa.exceptions.OperacaoNaoPermitidaException;
 import com.introducao.spring_jpa.model.Autor;
 import com.introducao.spring_jpa.repository.AutorRepository;
+import com.introducao.spring_jpa.repository.LivroRepository;
+import com.introducao.spring_jpa.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,15 +13,22 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository){
-        this.repository = repository;
-    }
+    // O COMPONENTE @RequiredArgConstructor na hora de compilar o código cria o construtor automaticamente
+//    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository){
+//        this.repository = repository;
+//        this.validator = validator;
+//        this.livroRepository = livroRepository;
+//    }
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -26,6 +37,7 @@ public class AutorService {
             throw new IllegalArgumentException("Para atualizar, é necessário que Autor já esteja salvo na base");
         }
 
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -34,6 +46,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException(autor.getNome() + " possui livros cadastrados");
+        }
         repository.delete(autor);
     }
 
@@ -53,6 +68,7 @@ public class AutorService {
         return repository.findAll();
     }
 
-
-
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existByAutor(autor);
+    }
 }
